feature(commit-types): add release as new type
Right now when you create a new release with semantic releases it creates a `chore` type commit. This commit includes all changes required for pushing a new version. But because `chore` also triggers a new patch version, this can be "looped". When someone creates a new release, and accidentally reruns the script, it will create yet another release.